requirements.txt
Flask
Flask-SQLAlchemy
APScheduler
requests
python-dotenv
authlib
cryptography
gunicorn
